+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Unreleased
 - Add :attr:`gidgethub.HTTPException.headers` attribute
   (`PR #227 <https://github.com/gidgethub/gidgethub/pull/227>`_)
 
+- Add a trailing ``/`` to *url* and remove any leading ``/`` in *base_url* for
+  :func:`gidgethub.sansio.format_url`
+
 5.4.0
 -----